Format class chaing and holding format definition
This will store any number format definition. Storing those into an
own class instances seems to be a little bit over the top but I'm
gonna use it as in indicator that the format definition has been
checked against a schema. Thus other class methods just need to check
whether a format defintion is an instance of this class to be sure
that the given format definition is valid.